# Conflicts:
#	Bundles/MultiCartPage/src/SprykerShop/Yves/MultiCartPage/Controller/MultiCartController.php

+++ Bundles/MultiCartPage/src/SprykerShop/Yves/MultiCartPage/Controller/MultiCartController.php
@@ -29,12 +29,15 @@ class MultiCartController extends AbstractController
      */
     protected const REMOVE_CART_ITEM_PERMISSION_PLUGIN_KEY = 'RemoveCartItemPermissionPlugin';
 
+    /**
+     * @uses \SprykerShop\Yves\CartPage\Plugin\Provider\CartControllerProvider::ROUTE_CART
+     */
+    protected const ROUTE_CART = 'cart';
+
     public const GLOSSARY_KEY_CART_UPDATED_SUCCESS = 'multi_cart_widget.cart.updated.success';
 
-    /**
-     * @deprecated Will be removed without replacement.
-     */
-    public const GLOSSARY_KEY_CART_WAS_DELETED = 'multi_cart_widget.cart.was-deleted-before';
+    protected const GLOSSARY_KEY_PERMISSION_FAILED = 'global.permission.failed';
+    protected const GLOSSARY_KEY_CART_UPDATED_ERROR = 'multi_cart_widget.cart.updated.error';
 
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
@@ -71,7 +74,7 @@ class MultiCartController extends AbstractController
                 ->createQuote($quoteTransfer);
 
             if ($quoteResponseTransfer->getIsSuccessful()) {
-                return $this->redirectResponseInternal(CartControllerProvider::ROUTE_CART);
+                return $this->redirectResponseInternal(static::ROUTE_CART);
             }
         }
 
@@ -110,6 +113,13 @@ class MultiCartController extends AbstractController
             ->handleRequest($request);
 
         $quoteTransfer = $quoteForm->getData();
+
+        if (!$quoteTransfer || !$this->canWriteQuote($quoteTransfer)) {
+            $this->addErrorMessage(static::GLOSSARY_KEY_PERMISSION_FAILED);
+
+            return $this->redirectResponseInternal(MultiCartPageControllerProvider::ROUTE_MULTI_CART_INDEX);
+        }
+
         if ($quoteForm->isSubmitted() && $quoteForm->isValid()) {
             $quoteResponseTransfer = $this->getFactory()
                 ->getMultiCartClient()
@@ -120,6 +130,10 @@ class MultiCartController extends AbstractController
 
                 return $this->redirectResponseInternal(MultiCartPageControllerProvider::ROUTE_MULTI_CART_INDEX);
             }
+
+            $this->addErrorMessage(static::GLOSSARY_KEY_CART_UPDATED_ERROR);
+
+            return $this->redirectResponseInternal(MultiCartPageControllerProvider::ROUTE_MULTI_CART_INDEX);
         }
 
         return [
@@ -141,7 +155,7 @@ class MultiCartController extends AbstractController
             ->getMultiCartClient()
             ->setDefaultQuote($quoteTransfer);
 
-        return $this->redirectResponseInternal(CartControllerProvider::ROUTE_CART);
+        return $this->redirectResponseInternal(static::ROUTE_CART);
     }
 
     /**
@@ -174,6 +188,12 @@ class MultiCartController extends AbstractController
     {
         $quoteTransfer = $this->findQuoteOrFail($idQuote);
 
+        if (!$this->isQuoteEditable($quoteTransfer)) {
+            $this->addErrorMessage(static::GLOSSARY_KEY_PERMISSION_FAILED);
+
+            return $this->redirectResponseInternal(MultiCartPageControllerProvider::ROUTE_MULTI_CART_INDEX);
+        }
+
         if (!$this->canRemoveCartItem($quoteTransfer)) {
             $this->addErrorMessage(static::MESSAGE_PERMISSION_FAILED);
 
@@ -188,7 +208,7 @@ class MultiCartController extends AbstractController
             $this->addSuccessMessage('multi_cart_page.cart_clear.success');
         }
 
-        return $this->redirectResponseInternal(CartControllerProvider::ROUTE_CART);
+        return $this->redirectResponseInternal(static::ROUTE_CART);
     }
 
     /**
@@ -199,6 +219,12 @@ class MultiCartController extends AbstractController
     public function deleteAction(int $idQuote)
     {
         $quoteTransfer = $this->findQuoteOrFail($idQuote);
+
+        if (!$this->canWriteQuote($quoteTransfer)) {
+            $this->addErrorMessage(static::GLOSSARY_KEY_PERMISSION_FAILED);
+
+            return $this->redirectResponseInternal(MultiCartPageControllerProvider::ROUTE_MULTI_CART_INDEX);
+        }
 
         $this->getFactory()->getMultiCartClient()->deleteQuote($quoteTransfer);
 
@@ -214,7 +240,7 @@ class MultiCartController extends AbstractController
 
         return $this->view(
             $response,
-            $this->getFactory()->getMultiCartListWidgetPlugins(),
+            [],
             '@MultiCartPage/views/cart/cart.twig'
         );
     }
@@ -242,10 +268,7 @@ class MultiCartController extends AbstractController
     {
         $viewData = $this->executeConfirmDeleteAction($idQuote);
 
-        $widgetPlugins = $this->getFactory()
-            ->getCartDeleteCompanyUsersListWidgetPlugins();
-
-        return $this->view($viewData, $widgetPlugins, '@MultiCartPage/views/cart-delete/cart-delete.twig');
+        return $this->view($viewData, [], '@MultiCartPage/views/cart-delete/cart-delete.twig');
     }
 
     /**
@@ -278,6 +301,29 @@ class MultiCartController extends AbstractController
         }
 
         return $quoteTransfer;
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\QuoteTransfer $quoteTransfer
+     *
+     * @return bool
+     */
+    protected function canWriteQuote(QuoteTransfer $quoteTransfer): bool
+    {
+        return $quoteTransfer->getCustomerReference() === $quoteTransfer->getCustomer()->getCustomerReference()
+            || $this->can('WriteSharedCartPermissionPlugin', $quoteTransfer->getIdQuote());
+    }
+
+    /**
+     * @param \Generated\Shared\Transfer\QuoteTransfer $quoteTransfer
+     *
+     * @return bool
+     */
+    protected function isQuoteEditable(QuoteTransfer $quoteTransfer): bool
+    {
+        return $this->getFactory()
+            ->getQuoteClient()
+            ->isQuoteEditable($quoteTransfer);
     }
 
     /**
